move classes to specific namespaces
