+++ src/cphase/calib/cameracalibration.cs
@@ -12,7 +12,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace I3DR.CPhase
+namespace I3DR.CPhase.Calib
 {
     //!  Camera Calibration class
     /*!
