+++ src/cphase/stereocamera/deimosstereocamera.cs
@@ -10,9 +10,9 @@
 
 using System;
 using System.Runtime.InteropServices;
-using I3DR.Phase;
+using I3DR.Phase.StereoCamera;
 
-namespace I3DR.CPhase
+namespace I3DR.CPhase.StereoCamera
 {
     //!  Deimos Stereo Camera class
     /*!
