+++ src/cphase/stereocamera/phobosstereocamera.cs
@@ -10,9 +10,9 @@
 
 using System;
 using System.Runtime.InteropServices;
-using I3DR.Phase;
+using I3DR.Phase.StereoCamera;
 
-namespace I3DR.CPhase
+namespace I3DR.CPhase.StereoCamera
 {
     //!  Phobos Stereo Camera class
     /*!
