+++ src/cphase/stereocamera/stereocamera.cs
@@ -12,9 +12,9 @@
 
 using System;
 using System.Runtime.InteropServices;
-using I3DR.Phase;
+using I3DR.Phase.StereoCamera;
 
-namespace I3DR.CPhase
+namespace I3DR.CPhase.StereoCamera
 {
     //!  Stereo Camera class
     /*!
