+++ src/cphase/stereomatcher/stereoi3drsgm.cs
@@ -11,7 +11,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace I3DR.CPhase
+namespace I3DR.CPhase.StereoMatcher
 {
     //!  Stereo I3DRSGM class
     /*!
