+++ src/phase/stereocamera/abstractstereocamera.cs
@@ -11,9 +11,9 @@
  */
 
 using System;
-using I3DR.CPhase;
+using I3DR.CPhase.StereoCamera;
 
-namespace I3DR.Phase
+namespace I3DR.Phase.StereoCamera
 {
 
     //!  Camera Read Result structure
