+++ src/phase/stereocamera/deimosstereocamera.cs
@@ -8,9 +8,9 @@
  * @details Capture data from I3DR's Deimos stereo camera
  */
 
-using I3DR.CPhase;
+using I3DR.CPhase.StereoCamera;
 
-namespace I3DR.Phase
+namespace I3DR.Phase.StereoCamera
 {
     //!  Deimos Stereo Camera class
     /*!
