+++ src/phase/stereocamera/pylonstereocamera.cs
@@ -8,9 +8,9 @@
  * @details Capture data from a stereo camera using Basler cameras.
  */
 
-using I3DR.CPhase;
+using I3DR.CPhase.StereoCamera;
 
-namespace I3DR.Phase
+namespace I3DR.Phase.StereoCamera
 {
     //!  Pylon Stereo Camera class
     /*!
