+++ src/phase/stereocamera/stereocamera.cs
@@ -10,9 +10,9 @@
  * used for any stereo camera. 
  */
 
-using I3DR.CPhase;
+using I3DR.CPhase.StereoCamera;
 
-namespace I3DR.Phase
+namespace I3DR.Phase.StereoCamera
 {
     //!  Stereo Camera class
     /*!
