+++ src/phase/stereomatcher/stereomatcher.cs
@@ -11,9 +11,9 @@
  */
 
 using System;
-using I3DR.CPhase;
+using I3DR.CPhase.StereoMatcher;
 
-namespace I3DR.Phase
+namespace I3DR.Phase.StereoMatcher
 {
     //!  Stereo Matcher class
     /*!
